+++ mycodo/databases/models/function.py
@@ -49,6 +49,9 @@ class ConditionalConditions(CRUDMixin, db.Model):
     # Output State
     output_id = db.Column(db.Text, default='')
 
+    # Controller
+    controller_id = db.Column(db.Text, default='')
+
     def __repr__(self):
         return "<{cls}(id={s.id})>".format(s=self, cls=self.__class__.__name__)